refactor service connection: consider distributed session ID

+++ cluster/agent.go
@@ -353,7 +353,7 @@ func (a *agent) writeMessageWS(m *message.Message) error {
 
 func (a *agent) notifySessionClosed(rpcClient *rpcClient, members []string) {
 	request := &clusterpb.SessionClosedRequest{
-		SessionId: a.session.ID(),
+		SessionId: int64(a.session.ID()),
 	}
 
 	for _, remote := range members {
