+++ examples/demo/tadpole/logic/manager.go
@@ -21,7 +21,7 @@ func NewManager() *Manager {
 // Login handler was used to guest login
 func (m *Manager) Login(s *session.Session, msg *protocol.JoyLoginRequest) error {
 	log.Println(msg)
-	id := s.ID()
+	id := int64(s.ID())
 	s.Bind(id)
 	return s.Response(protocol.LoginResponse{
 		Status: protocol.LoginStatusSucc,
